+++ process/similarity.py
@@ -4,17 +4,19 @@ import psycopg
 import os
 from urllib.parse import urlparse
 from time import sleep, time
-import gc
 
-def process_similarity():
-  with psycopg.connect(os.environ["DB_URI"]) as conn:
-    with conn.cursor() as cur:
-      cur.execute("SELECT (id, readme_url, repo_url) FROM ships")
-      ships = cur.fetchall()
+def process_similarity(pre_ships):
+  if pre_ships != None:
+    ships = list(map(lambda s: ((s["id"], s["fields"]["readme_url"], s["fields"]["repo_url"]),), pre_ships))
+  else:
+    with psycopg.connect(os.environ["DB_URI"]) as conn:
+      with conn.cursor() as cur:
+        cur.execute("SELECT (id, readme_url, repo_url) FROM ships")
+        ships = cur.fetchall()
 
-      cur.close()
-  
-  conn.close()
+        cur.close()
+    
+    conn.close()
 
   if len(ships) == 0:
     print("Ships not downloaded.")
@@ -81,52 +83,51 @@ def process_similarity():
   print("Calculating similarity...")
 
   print("Building similarity pairs...")
-  with psycopg.connect(os.environ["DB_URI"]) as conn:
-    with conn.cursor() as cur:
-      cur.execute("DELETE FROM similarity")
 
-      insertArgs = []
-      for x in range(len(id_list)):
-        for y in range(len(id_list)):
-          if x == y:
-            continue
-            
-          ai_value = normalized_list[x].similarity(normalized_list[y])
-
-          lang_similarity = 0
-          if languages[x] != None and languages[y] != None:
-            if len(languages[x].keys()) != 0 and len(languages[y].keys()) != 0:
-              top_language_x = None
-              for lang in languages[x]:
-                if top_language_x == None:
-                  top_language_x = lang
-                elif languages[x][lang] > languages[x][top_language_x]:
-                  top_language_x = lang
-              
-              top_language_y = None
-              for lang in languages[y]:
-                if top_language_y == None:
-                  top_language_y = lang
-                elif languages[y][lang] > languages[y][top_language_y]:
-                  top_language_y = lang
-              
-              lang_similarity = int(top_language_y == top_language_x)
-          
-          insertArgs.append((id_list[x], id_list[y], ai_value, lang_similarity))
-    
-      updateArgs = []
-      for ship in filtered_ships:
-        updateArgs.append((ship,))
+  insertArgs = []
+  for x in range(len(id_list)):
+    for y in range(len(id_list)):
+      if x == y:
+        continue
         
-      cur.executemany("INSERT INTO similarity (shipa, shipb, nlp_value, lang_value) VALUES (%s, %s, %s, %s)", insertArgs)
-      cur.executemany("UPDATE ships SET filtered = true WHERE id = %s", updateArgs)
+      ai_value = normalized_list[x].similarity(normalized_list[y])
 
-      cur.close()
-    
-    conn.commit()
-    conn.close()
+      lang_similarity = 0
+      if languages[x] != None and languages[y] != None:
+        if len(languages[x].keys()) != 0 and len(languages[y].keys()) != 0:
+          top_language_x = None
+          for lang in languages[x]:
+            if top_language_x == None:
+              top_language_x = lang
+            elif languages[x][lang] > languages[x][top_language_x]:
+              top_language_x = lang
+          
+          top_language_y = None
+          for lang in languages[y]:
+            if top_language_y == None:
+              top_language_y = lang
+            elif languages[y][lang] > languages[y][top_language_y]:
+              top_language_y = lang
+          
+          lang_similarity = int(top_language_y == top_language_x)
+      
+      insertArgs.append((id_list[x], id_list[y], ai_value, lang_similarity))
 
-  del nlp
-  gc.collect()
+  updateArgs = []
+  for ship in filtered_ships:
+    updateArgs.append((ship,))
 
   print("Done with similarity")
+  if pre_ships != None:
+    return (insertArgs, list(filtered_ships.keys()))
+  else:
+    with psycopg.connect(os.environ["DB_URI"]) as conn:
+      with conn.cursor() as cur:
+        cur.execute("DELETE FROM similarity")  
+        cur.executemany("INSERT INTO similarity (shipa, shipb, nlp_value, lang_value) VALUES (%s, %s, %s, %s)", insertArgs)
+        cur.executemany("UPDATE ships SET filtered = true WHERE id = %s", updateArgs)
+
+        cur.close()
+      
+      conn.commit()
+      conn.close()
